requirements.txt
streamlit==1.44.1
python-dotenv==1.1.0
pandas 